Share the invariant globalization check between the Unicode and Xunit packages
