+++ src/Meziantou.Xunit/Meziantou.Xunit.csproj
@@ -10,6 +10,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="../../common/GlobalizationHelper.cs" Link="GlobalizationHelper.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="xunit.v3.extensibility.core" Version="4.0.0" />
   </ItemGroup>
